Updated TyrusSupport to correctly propagate query params from webserver to Tyrus. Moved UriComponent over to common/http for sharing purposes. Updated tests to verify changes.
