+++ common/http/src/test/java/io/helidon/common/http/UriComponentTest.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2017, 2021 Oracle and/or its affiliates.
+ * Copyright (c) 2022 Oracle and/or its affiliates.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,13 +14,11 @@
  * limitations under the License.
  */
 
-package io.helidon.webserver;
+package io.helidon.common.http;
 
 import java.net.URI;
 import java.net.URLEncoder;
 import java.util.Optional;
-
-import io.helidon.common.http.Parameters;
 
 import org.junit.jupiter.api.Test;
