+++ webserver/tyrus/src/main/java/module-info.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2020 Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2020, 2022 Oracle and/or its affiliates.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -26,6 +26,7 @@ module io.helidon.webserver.tyrus {
     requires io.helidon.common.context;
     requires io.helidon.common.mapper;
     requires io.helidon.common.reactive;
+    requires io.helidon.common.http;
 
     requires tyrus.core;
     requires tyrus.server;
